backend/requirements.txt
# =========================
# Core Framework
# =========================
fastapi==0.123.5
uvicorn==0.38.0

# =========================
# Database
# =========================
SQLAlchemy==2.0.44
pymysql==1.1.1

# =========================
# Auth / Security
# =========================
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4

# =========================
# Validation
# =========================
pydantic[email]==2.12.5
python-multipart==0.0.9

# =========================
# Utils
# =========================
python-dotenv==1.2.1

# =========================
# Data & Visualization  ← ⭐ 분석/리포트용 추가 구역
# =========================
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.1
